mac80211: Add mesh sync variables and netlink interface
This patch adds the defines and enums to netlink, which are needed for
the extensible synchronization interface and the default neighbor offset
synchronization.  The mesh sync method is loaded on mesh start from
mesh_setup entries. This value is currently not/unknown accessible to
netlink.  The maximum number of neighbors for neighbor offset sync is
placed in mesh_config and thus may be changed while the mesh is up. A
default value of 50 is defined according to the standards proposal.

The hardcoded neighbor offset entry of zero in mesh_sp_id is removed.

Signed-off-by:  Marco Porsch <marco.porsch@etit.tu-chemnitz.de>
Reviewed-by:  Ashok Nagarajan <ashok@cozybit.com>
Reviewed-by:  Javier Cardona <javier@cozybit.com>