mac80211: assign aid when adding mesh peers
When allocating a mesh peer from the mac80211 MPM, assign a random AID.
Use this value as the llid if initiating the peering, or the peer's
value if responding to a peering open frame.

Needed for some firmware which requires a valid AID when adding a
station, and future mesh power save implementation.