ocfs2: skip locks in the blocked list
  A parallel umount on 4 nodes triggered a bug in dlm_process_recovery_date(). Here’s the situation:
  Receiving MIG_LOCKRES message, A node processes the locks in migratable lockres. It copys lvb from migratable lockres when processing the first valid lock.
If there is a lock in the blocked list with the EX level, it triggers the BUG. Since valid lvbs are set when locks are granted with EX or PR levels, locks in
the blocked list cannot have valid lvbs. Therefore I think we should skip the locks in the blocked list.

Signed-off-by: Xuejiufei <xuejiufei@huawei.com>
Signed-off-by: Joel Becker <jlbec@evilplan.org>