RMoM: squashed work
This commit rebases all inital RMoM work up until 10/4/2012 on top of
linux 3.5. It's just a diff from
fc79118fe2b545dadb3812fb92491e409764424c~1 to
939b99d9263ee50d024455d668e41083134b7f61 I had to apply manually.

I tried retaining the commits, but for some reason there were
too many conflicts. We will have to rework the patches before submitting
upstream anyway, so this shouldn't be too big of a deal.

A nice outcome of this rebase is generating a compat-drivers now goes
smoothly! We should keep open80211s.git/master pointing to the latest
linux release in the future, so compat-drivers generation is easy for
master and feature branches.