ARM: gemini: remove platform support
The gemini platform has not seen any updates from its
maintainer since 2010, and has no defconfig file, which
has lead to bitrot in a number of places:

* System restart has been broken for a long time despite
  repeated attempts from Russell to get the obvious
  fix at least acknowledged by someone who can test it.
* The platform still uses mach/gpio.h, which has been
  deprecated for some time, and since Kconfig does not
  select NEED_MACH_GPIO_H, it actually failed to build
  because of this.
* A change to asm/io.h introduced a number of compiler
  warnings for gemini specific code that confuses
  physical and virtual addresses.

This removes gemini support entirely under the assumption
that nobody uses it on mainline kernels. Of course, if
there are still active users out there, they should let
us know so we can put support back in and get it better
tested.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>