mac80211: always forward on all interfaces
This is enough to make ping work across the bridge.  Previously,
forwarded unicast frames that had a destination on the other
channel would be resent on the wrong interface.  We should really
use the proper interface for unicast frames instead of blindly
rebroadcasting on all, but this at least shows the mechanism works.

Signed-off-by: Bob Copeland <me@bobcopeland.com>