mac80211: make RMC global
The RMC was previously per-sdata, but it is sufficient for
the linux host to receive a multicast frame once.

Now that multiple interfaces can share the same mesh data
forwarding duties we may inadvertently create a multicast
routing loop and have a corner case where a multicast
frame is retransmitted a few times too many.

Therefore it makes sense to make the RMC global to
mac80211.

Signed-off-by: Thomas Pedersen <thomas@cozybit.com>