mac80211: handle headroom for multi-if mesh
When an MBSS is shared across multiple interfaces, the
headroom requirements may differ. Handle this by
indicating the max headroom of all PHYs in the mesh to
each member netdev, so at TX time sufficient headroom
space is guaranteed to be available regardless of whether
an interface switch takes place.

In the forwarding path, just allocate up to max mbss
headroom before looking up the outgoing interface, which
may require a reallocation anyway.

Signed-off-by: Thomas Pedersen <thomas@cozybit.com>