mwl8787: rewrite cmd handling for hw specific header
We need to be able to prepend a hardware-specific header
for certain busses (sdio) so allocate with some headroom.
Still need to do something about block size needed for sdio
transfers, but this gets closer to avoiding a copy in the
send_cmd.

Also simplify the structures a bit and change some defines
to enums.

Signed-off-by: Bob Copeland <me@bobcopeland.com>