mwl8787: allow setting tx power
According to firmware code, 30 dBm is the default tx power for
8787.  However, we were never setting it to anything in the
first place, and the firmware apparently doesn't load in the
defaults until you call the set power command.  Just go ahead
and set it to 30 dBm for all rates and allow configuration through
netlink to further restrict it.

This increases my throughput at a few meters away to from 5 to
15 mbps.

Signed-off-by: Bob Copeland <me@bobcopeland.com>