FS-Cache: Fix handling of an attempt to store a page that is now beyond EOF
Fix the handling of an attempt to store a page that is now beyond EOF.  This
may happen, for example, if the page got pushed for storage before the netfs
file got truncated on the server.  In such a case, we should just remove the
excessive pages from the cookie->stores radix tree and wake up the waiter.

This can be seen in /proc/fs/fscache/stats on this line:

	Stores : ops=350 run=1895 pgs=1545 rxd=1727 olm=9

where olm=N has N > 0.

Reported-by: Milosz Tanski <milosz@adfin.com>
Signed-off-by: David Howells <dhowells@redhat.com>
Acked-by: Milosz Tanski <milosz@adfin.com>