ARM: make mach-xyz/Makefile.boot optional for !ARCH_MULTIPLATFORM
Makefile.boot is supposed to define zreladdr-y, params_phys-y and
initrd_phys-y. The first one is only needed in the absence of AUTO_ZRELADDR
and when building a zImage. The latters are only needed for bootp stuff.
So ignore errors on including Makefile.boot and error out in the targets
that need the corresponding settings.

This makes it unnecessary to create dummy Makefile.boot files.

Signed-off-by: Uwe Kleine-König <u.kleine-koenig@pengutronix.de>