sched: replace INIT_COMPLETION with reinit_completion
For the casual device driver writer, it is hard to remember when to use
init_completion (to init a completion structure) or INIT_COMPLETION (to
*reinit* a completion structure).  Furthermore, while all other completion
functions exepct a pointer as a parameter, INIT_COMPLETION does not.  To
make it easier to remember which function to use and to make code more
readable, introduce a new inline function with the proper name and
consistent argument type.  Update the kernel-doc for init_completion while
we are here.

Signed-off-by: Wolfram Sang <wsa@the-dreams.de>
Acked-by: Linus Walleij <linus.walleij@linaro.org> (personally at LCE13)
Cc: Ingo Molnar <mingo@kernel.org>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>