mac80211: mesh: only get tsf if we need it
The local TSF is used to compute the timing offset between
mesh peers on beacon reception.  However, asking the card for
the TSF is not very accurate, so we prefer to use rx->mactime.
In the latter case, calling drv_get_tsf() just adds more delay
into the RX path, so skip it if we can.

Signed-off-by: Bob Copeland <me@bobcopeland.com>